Set data type for `NULL` params
